Refactor: create rest_connector module
The rest connector module moves most of the functionality of http
connection out of the client and into a module. The goal is to simplify
method calls and DRY up the code.

Changes include replacing all calls to send_request with calls to either
'get' or 'post'. process request now returns the "data" portion of the
JSON response from the server, as every method was retrieving that
separately.

removed some untested code.
