+++ features/step_definitions/step_helpers.rb
@@ -66,7 +66,7 @@ def new_client_from_stored_values
 end
 
 def get_token_from_file
-  token = JSON.parse(File.read(BitPay::TOKEN_FILE_PATH))['data'][0]
+  token = JSON.parse(File.read(BitPay::TOKEN_FILE_PATH))[0]
   {token['facade'] => token['token']}
 end
